Make Commitment an associated type.
Usually, it will be instantiated using HashCommitment<Self>, which
this crate provides. However, this gives flexibility to delegate
through an intermediate type, if desired. For example, in Espresso,
Both ValidatorState and LedgerCommitmentOpening can have
`type Commitment = LedgerStateCommitment`.

We provide a type alias `type Commitment<T> = <T as Committable>::Commitment`
for convenience and compatibility.